# bugfix: fix the rope correctness issue introduced in #609